// packages/module/blockchain-cardano-ui/src/components/ActivityDetails/ProposalProcedures/NoConfidenceAction/types.ts
import type * as ActionIdTypes from '../components/ActionIdTypes';
import type * as ProcedureTypes from '../components/ProcedureTypes';
import type * as TxDetailsTypes from '../components/ProposalProcedureTransactionDetailsTypes';

export interface Data {
  procedure: ProcedureTypes.Procedure;
  actionId?: ActionIdTypes.Data;
  txDetails: TxDetailsTypes.TxDetails;
}

export interface Translations {
  procedure: ProcedureTypes.Translations;
  actionId?: ActionIdTypes.Translations;
  txDetails: TxDetailsTypes.Translations;
}